refactor: move logic to services, req/res to controller

+++ src/post/album/album.service.ts
@@ -1,49 +1,49 @@
-import { HttpStatus, Injectable } from '@nestjs/common';
-import { isUUID } from 'class-validator';
+import { Injectable } from '@nestjs/common';
 import { randomUUID } from 'crypto';
 import { Album } from 'src/common/interfaces';
 import { DataService } from 'src/common/services';
-import { isAlbumData } from 'src/common/utils';
 
 @Injectable()
 export class AlbumService {
   constructor(private dataService: DataService) {}
-  create(createArtistDto: Album) {
-    let s = HttpStatus.CREATED;
-    if (!isAlbumData(createArtistDto)) s = HttpStatus.BAD_REQUEST;
+  create(data: Album) {
     const dto = {
       id: randomUUID({ disableEntropyCache: true }),
-      ...createArtistDto,
+      ...data,
     };
-    return { status: s, body: this.dataService.createAlbum(dto) };
+    this.dataService.createAlbum(dto);
+    return dto;
   }
 
   findAll() {
-    return { status: HttpStatus.OK, body: this.dataService.getAllAlbums() };
+    return this.dataService.getAllAlbums();
   }
 
   findOne(id: string) {
-    let s = HttpStatus.OK;
-    if (!isUUID(id, '4')) s = HttpStatus.BAD_REQUEST;
-    const album = this.dataService.getAlbum(id);
-    if (!album && s === HttpStatus.OK) s = HttpStatus.NOT_FOUND;
-    return { status: s, body: album };
+    return this.findAll().find((el) => el.id === id);
   }
 
-  update(id: string, updateAlbumDto: Album) {
-    let s = HttpStatus.OK;
-    if (!isUUID(id, '4') || !isAlbumData(updateAlbumDto))
-      s = HttpStatus.BAD_REQUEST;
-    const album = this.dataService.updateAlbum(id, updateAlbumDto);
-    if (!album && s === HttpStatus.OK) s = HttpStatus.NOT_FOUND;
-    return { status: s, body: album };
+  update(id: string, dto: Album) {
+    const data = this.dataService.getAllAlbums();
+    const index = data.findIndex((el) => el.id === id);
+    if (index < 0) return false;
+    const updatedDto = {
+      ...data[index],
+      ...dto,
+    };
+    this.dataService.updateAlbum(index, updatedDto);
+    return updatedDto;
   }
 
   remove(id: string) {
-    let s = HttpStatus.NO_CONTENT;
-    if (!isUUID(id, '4')) s = HttpStatus.BAD_REQUEST;
-    const album = this.dataService.deleteAlbum(id);
-    if (!album && s === HttpStatus.NO_CONTENT) s = HttpStatus.NOT_FOUND;
-    return { status: s, body: album };
+    this.dataService.getAllTracks().forEach((track, index) => {
+      if (track.albumId === id) {
+        this.dataService.updateTrack(index, { ...track, albumId: null });
+      }
+    });
+    const data = this.findOne(id);
+    if (!data) return false;
+    this.dataService.deleteAlbum(id);
+    return id;
   }
 }
