+++ src/post/favorites/track/track.service.ts
@@ -7,25 +7,19 @@ export class FavTrackService {
   constructor(private dataService: DataService) {}
 
   create(id: string) {
-    let s = HttpStatus.CREATED;
-    if (!isUUID(id, '4')) s = HttpStatus.BAD_REQUEST;
-    const createResponse = this.dataService.addTrackToFav(id);
-    if (!createResponse && s === HttpStatus.CREATED)
-      s = HttpStatus.UNPROCESSABLE_ENTITY;
-    return { status: s, body: createResponse };
+    const isExists = this.dataService.getAllTracks().find((el) => el.id === id);
+    if (!isExists) return false;
+    this.dataService.addTrackToFav(id);
+    return id;
   }
 
   findAll() {
-    const allTrackFavs = this.dataService.getAllTrackFavs();
-    return { status: HttpStatus.OK, body: allTrackFavs };
+    return this.dataService.getAllTrackFavs();
   }
 
   remove(id: string) {
-    let s = HttpStatus.NO_CONTENT;
-    if (!isUUID(id, '4')) s = HttpStatus.BAD_REQUEST;
-    const deleteResponse = this.dataService.deleteTrackFav(id);
-    if (!deleteResponse && s === HttpStatus.NO_CONTENT)
-      s = HttpStatus.UNPROCESSABLE_ENTITY;
-    return { status: s, body: deleteResponse };
+    if (!this.findAll().includes(id)) return false;
+    this.dataService.deleteTrackFav(id);
+    return id;
   }
 }
